not working

+++ Programs/Program3/pdu.h
@@ -29,8 +29,8 @@ uint32_t getHSeqNum(pduPacket *pduBuffer);
 uint8_t getFlag(pduPacket *pduBuffer);
 void getFromFileName(pduPacket *pduBuffer, int pduLen, char *fileName);
 void setFlag(pduPacket *pduBuffer, int pduLen, uint8_t flag);
-uint16_t getWindowSizeFromPDU(pduPacket *pduBuffer);
-uint32_t getPayloadSizeFromPDU(pduPacket *pduBuffer);
+uint32_t getWindowSizeFromPDU(pduPacket *pduBuffer);
+uint16_t getPayloadSizeFromPDU(pduPacket *pduBuffer);
